+++ src/bloqade/ir/pulse.py
@@ -2,7 +2,7 @@ from .scalar import Interval
 from .field import Field
 from typing import List
 from pydantic.dataclasses import dataclass
-from .print import Printer
+from .tree_print import Printer
 
 
 __all__ = [
